pdf-qabot/requirements.txt
fastapi
uvicorn
pypdf
langchain
sentence-transformers
faiss-cpu
python-dotenv
